TST: unit test restructure
Restructured unit tests and implemented pytest in place of nose.
